+++ model/user.go
@@ -1,8 +1,10 @@
 package model
 
 type User struct {
-	Id    uint
+	Id    int64
 	Email string
+	// CreatedAt time.Time
+	// UpdatedAt time.Time
 	// Notes
 	// Subscriptions
 	// Collections
